rename variables and roles file.
