+++ src/config/index.js
@@ -1,1 +1,1 @@
-export { default as Roles } from './roles';
+export { default as Roles } from './Roles';
