+++ src/routes/MapAllowedRoutes.js
@@ -1,11 +1,10 @@
 import React, { memo } from 'react';
 import { Switch, Route, useRouteMatch } from 'react-router-dom';
-import { isArrayWithLength } from 'utils';
 import { NotFound } from 'components/common';
 
 /*
-* This is the route utility component used for
-* generating child routes it only requires routes and basePath
+* This is the route utility component used for generating
+* routes and child routes it only requires routes array and basePath
 */
 function MapAllowedRoutes({routes, basePath}) {
 	const match = useRouteMatch(basePath);
@@ -18,18 +17,21 @@ function MapAllowedRoutes({routes, basePath}) {
 				* just make sure that rest object only contain props that supported by react-router's route component
 				* you may find props list here https://reactrouter.com/web/api/Route
 				*/
-				const { path, component: Component, children, title, permission, ...rest } = route;
+				const {
+					path,
+					component: Component,
+					children,
+					title,
+					permission,
+					...rest
+				} = route;
 				return (
 					<Route
 						{...rest}
 						key={path}
 						path={`${match.path}${path}`}
 					>
-						{
-							isArrayWithLength(children) ?
-								<Component children={children} /> :
-								<Component />
-						}
+						<Component children={children} />
 					</Route>
 				)
 			})}
